# documentation/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info":
  {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers":
  [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths":
  {
    "/api/teams/update":
    {
      "put":
      {
        "tags":
        [
          "team-controller"
        ],
        "summary": "Update team",
        "operationId": "updateTeam",
        "requestBody":
        {
          "content":
          {
            "application/json":
            {
              "schema":
              {
                "type": "string",
                "example":
                {
                  "name": "My awesome team",
                  "level": 2,
                  "score": 15,
                  "lives": 4,
                  "active": true
                }
              }
            }
          },
          "required": true
        },
        "responses":
        {
          "201":
          {
            "description": "Team updated",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          },
          "404":
          {
            "description": "Team not found",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/update-player-name/{email}":
    {
      "put":
      {
        "tags":
        [
          "player-controller"
        ],
        "summary": "Find player by email and update name",
        "operationId": "updatePlayerName",
        "parameters":
        [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema":
            {
              "type": "string"
            }
          }
        ],
        "requestBody":
        {
          "content":
          {
            "application/json":
            {
              "schema":
              {
                "type": "string",
                "example": "New player name"
              }
            }
          },
          "required": true
        },
        "responses":
        {
          "200":
          {
            "description": "Player updated",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "400":
          {
            "description": "Bad request",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/teams/join":
    {
      "post":
      {
        "tags":
        [
          "team-controller"
        ],
        "summary": "Join team",
        "operationId": "joinTeam",
        "requestBody":
        {
          "content":
          {
            "application/json":
            {
              "schema":
              {
                "type": "string",
                "example":
                {
                  "teamName": "My awesome team",
                  "playerEmail": "test@test.de"
                }
              }
            }
          },
          "required": true
        },
        "responses":
        {
          "201":
          {
            "description": "Team joined",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          },
          "400":
          {
            "description": "Bad request",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/teams/create":
    {
      "post":
      {
        "tags":
        [
          "team-controller"
        ],
        "summary": "Create team",
        "operationId": "createTeam",
        "requestBody":
        {
          "description": "Name of the team",
          "content":
          {
            "application/json":
            {
              "schema":
              {
                "type": "string",
                "example": "My awesome team"
              }
            }
          },
          "required": true
        },
        "responses":
        {
          "201":
          {
            "description": "Team created",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          },
          "400":
          {
            "description": "Bad request",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/register":
    {
      "post":
      {
        "tags":
        [
          "player-controller"
        ],
        "summary": "Create new player",
        "operationId": "registerPlayer",
        "requestBody":
        {
          "content":
          {
            "application/json":
            {
              "schema":
              {
                "type": "string",
                "example":
                {
                  "name": "Player 1",
                  "email": "test@test.de",
                  "password": "pw123"
                }
              }
            }
          },
          "required": true
        },
        "responses":
        {
          "400":
          {
            "description": "Bad request",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "object"
                }
              }
            }
          },
          "201":
          {
            "description": "Player created",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/login":
    {
      "post":
      {
        "tags":
        [
          "player-controller"
        ],
        "summary": "Find player by email and password",
        "operationId": "loginPlayer",
        "requestBody":
        {
          "content":
          {
            "application/json":
            {
              "schema":
              {
                "type": "string",
                "example":
                {
                  "email": "test@test.de",
                  "password": "pw123"
                }
              }
            }
          },
          "required": true
        },
        "responses":
        {
          "200":
          {
            "description": "Player found",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "400":
          {
            "description": "Bad request",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "string"
                }
              }
            }
          },
          "404":
          {
            "description": "Player not found",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/teams":
    {
      "get":
      {
        "tags":
        [
          "team-controller"
        ],
        "summary": "Find all teams",
        "operationId": "getTeams",
        "responses":
        {
          "200":
          {
            "description": "OK",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "array",
                  "items":
                  {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/teams/{name}":
    {
      "get":
      {
        "tags":
        [
          "team-controller"
        ],
        "summary": "Find team by name",
        "operationId": "getTeamByName",
        "parameters":
        [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema":
            {
              "type": "string"
            }
          }
        ],
        "responses":
        {
          "200":
          {
            "description": "Team found",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          },
          "404":
          {
            "description": "Team not found",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/teams/player/{email}":
    {
      "get":
      {
        "tags":
        [
          "team-controller"
        ],
        "summary": "Find teams for player",
        "operationId": "getTeamByPlayer",
        "parameters":
        [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema":
            {
              "type": "string"
            }
          }
        ],
        "responses":
        {
          "404":
          {
            "description": "Team not found",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "object"
                }
              }
            }
          },
          "200":
          {
            "description": "Team for player found",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "array",
                  "items":
                  {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/teams/player/{email}/active":
    {
      "get":
      {
        "tags":
        [
          "team-controller"
        ],
        "summary": "Find active teams for player",
        "operationId": "getActiveTeamByPlayer",
        "parameters":
        [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema":
            {
              "type": "string"
            }
          }
        ],
        "responses":
        {
          "200":
          {
            "description": "Active team for player found",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "array",
                  "items":
                  {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          },
          "404":
          {
            "description": "Team not found",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/players":
    {
      "get":
      {
        "tags":
        [
          "player-controller"
        ],
        "summary": "Find all players",
        "operationId": "getAllPlayers",
        "responses":
        {
          "200":
          {
            "description": "OK",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "array",
                  "items":
                  {
                    "$ref": "#/components/schemas/Player"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/players/delete-all":
    {
      "delete":
      {
        "tags":
        [
          "player-controller"
        ],
        "summary": "Delete all players",
        "operationId": "deletePlayers",
        "responses":
        {
          "200":
          {
            "description": "All players deleted",
            "content":
            {
              "*/*":
              {
                "schema":
                {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components":
  {
    "schemas":
    {
      "Player":
      {
        "type": "object",
        "properties":
        {
          "id":
          {
            "type": "integer",
            "format": "int64"
          },
          "name":
          {
            "type": "string"
          },
          "email":
          {
            "type": "string"
          },
          "password":
          {
            "type": "string"
          }
        }
      },
      "Team":
      {
        "type": "object",
        "properties":
        {
          "id":
          {
            "type": "integer",
            "format": "int64"
          },
          "name":
          {
            "type": "string"
          },
          "level":
          {
            "type": "integer",
            "format": "int32"
          },
          "score":
          {
            "type": "integer",
            "format": "int32"
          },
          "lives":
          {
            "type": "integer",
            "format": "int32"
          },
          "active":
          {
            "type": "boolean"
          },
          "playerLeft":
          {
            "$ref": "#/components/schemas/Player"
          },
          "playerRight":
          {
            "$ref": "#/components/schemas/Player"
          }
        }
      }
    }
  }
}